Minor edit of README.rst
I made some minor changes to the language in the readme and ensured all
lines were under 80 characters long.. Most importantly, I changed all
of the links from `sharat87/ti` to go to the forked project at
`tbekolay/ti` on GitHub. This way, users like me would not accidentally
create new GitHub issues on the unmaintained project, or accidentally
download the out-of-date source when installing `ti` for the first
time.